Added a new backend: mount
This is quite rought and brutal, it read the fstab file
to see the availabel mount points and the /proc/mounts
files to see if they are mounted or not.

atm only looks for fns and cifs filesystem, as all the
others should be yet managed by udisks. Works quite well
with my nfs shares.

TBD: actually mount/unmount using the mount command